requirements.txt
asgiref==3.7.2
Django==4.2.4
django-mathfilters==1.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
